Move WalletState to separate file
