+++ lib/home.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 import 'package:bitcoin_ui/bitcoin_ui.dart';
 import 'package:donationwallet/load_wallet.dart';
 import 'package:donationwallet/main.dart';
+import 'package:donationwallet/states/wallet_state.dart';
 import 'package:donationwallet/tx_history.dart';
 import 'package:donationwallet/wallet.dart';
 import 'package:donationwallet/settings.dart';
