+++ lib/outputs.dart
@@ -1,4 +1,5 @@
 import 'package:donationwallet/main.dart';
+import 'package:donationwallet/states/wallet_state.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:donationwallet/rust/api/simple.dart';
