+++ lib/settings.dart
@@ -3,6 +3,7 @@ import 'package:donationwallet/rust/api/simple.dart';
 import 'package:donationwallet/global_functions.dart';
 import 'package:donationwallet/main.dart';
 import 'package:donationwallet/home.dart';
+import 'package:donationwallet/states/wallet_state.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
